requirements.txt
pyside6-essentials>=6.7
pyodbc>=5.1
sqlglot>=27.0
Pillow>=10.0
psutil>=6.0
